% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_list.R
\name{occ_list}
\alias{occ_list}
\title{Get a species list}
\usage{
occ_list(query = NULL, from = "gbif", limit = 500, geometry = NULL)
}
\description{
Get a species list
}
\examples{
\dontrun{
occ_list(geometry = "")
}
}

