% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_list_gbif.R
\name{occ_list_gbif}
\alias{occ_list_gbif}
\title{Get a species list}
\usage{
occ_list_gbif(country = NULL, publishingCountry = NULL, typeStatus = NULL,
  lastInterpreted = NULL, continent = NULL, geometry = NULL,
  recordedBy = NULL, basisOfRecord = NULL, datasetKey = NULL,
  eventDate = NULL, year = NULL, month = NULL, decimalLatitude = NULL,
  decimalLongitude = NULL, elevation = NULL, depth = NULL,
  institutionCode = NULL, collectionCode = NULL, search = NULL,
  limit = 500, start = 0)
}
\description{
Get a species list
}
\examples{
\dontrun{
geometry <- "POLYGON((8.98 48.05,15.66 48.05,15.66 45.40,8.98 45.40,8.98 48.05))"
occ_list_gbif(geometry = geometry, limit = 20)

# search by country
(res <- occ_list_gbif(country = 'US', limit = 20))
(res <- occ_list_gbif(country = 'CA', limit = 20))
}
}

