% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranker.R
\name{ranker}
\alias{ranker}
\title{Clean a species list}
\usage{
ranker(x, by = "species")
}
\arguments{
\item{x}{input}

\item{by}{(character) By something, only option now: "species"}
}
\description{
Clean a species list
}
\examples{
\dontrun{
(res <- occ_list_gbif(country = 'US', limit = 20))
ranker(res, "species")
}
}
