% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_list.R
\name{spp_list}
\alias{spp_list}
\title{Get a species list}
\usage{
spp_list(query = NULL, from = "gbif", limit = 500, geometry = NULL)
}
\description{
Get a species list
}
\examples{
\dontrun{
geometry <- "POLYGON((8.98 48.05,15.66 48.05,15.66 45.40,8.98 45.40,8.98 48.05))"

# gbif
spp_list(geometry = geometry)

# gbif, obis, and bison
res <- spp_list(geometry = geometry, from = c('gbif', 'obis', 'ebird'), limit = 20)
res$gbif
res$obis
res$ebird
}
}

